api/pyproject.toml
[tool.poetry]
name = "summsup"
version = "0.1.0"
description = ""
authors = ["Victor Resende <victoresende19@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.4"
uvicorn = "^0.32.0"
python-multipart = "^0.0.17"
langchain = "^0.3.7"
openai = "^1.54.3"
pillow = "^11.0.0"
langchain-community = "^0.3.5"
pypdf2 = "^3.0.1"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
